Server v2
This commit takes HL to v1.3.0! It changes how HL's server API works, primarily allowing for a Handler to be written outside of the crate where the trait is written. This allows for closed source servers - something we're doing at Valera with Atlas.

The server now has no wrapper in the `rpc` macro, simplifying it. The server is now blocking and a little less complex by inputting the factory when you create the server instead of using generics.